feat: add `TextArea` and `FileUpload` component
This adds two new components and fixes the name of an icon.

An expanding `TextArea` component is added that has changed many of the
elements by changing the fixed `height` of Tailwind's `h-14` (3.5 rem)
to a `min-height` of `h-14`. Combining this with `height: auto` allows
for expanding elements, such as a `TextArea`, while still using the `FormControl`
element underneath.

A `FileUpload` component is added that allows the end user to add files.
As the native element doesn't allow one to remove files, due to the use
of `ReadonlyArray<File>`, the files are pushed to an internal `Array`,
which is mutable.

The icon `arrow-circle-down` has been renamed to match
`circle-arrow-down`, as is exported by Figma.